requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
anthropic==0.25.6
anyio==4.3.0
argcomplete==3.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arize==7.15.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
azure-core==1.30.1
azure-mgmt-core==1.4.0
azure-storage-blob==12.19.1
Babel==2.14.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.7.0
blis==0.7.11
build==1.2.1
CacheControl==0.14.0
cachetools==5.3.3
catalogue==2.0.10
cattrs==23.2.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.5.0
cleo==2.1.0
click==8.1.7
cloudpathlib==0.16.0
coloredlogs==15.0.1
comm==0.2.2
confection==0.1.4
context-python==0.13.0
contextvars==2.4
crashtest==0.4.1
cryptography==42.0.5
cymem==2.0.8
dataclasses-json==0.6.4
datasets==2.19.0
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.8
dirtyjson==1.0.8
distlib==0.3.8
distro==1.9.0
dnspython==2.6.1
docstring_parser==0.16
dulwich==0.21.7
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.7.1/en_core_web_md-3.7.1-py3-none-any.whl#sha256=6a0f857a2b4d219c6fa17d455f82430b365bf53171a2d919b9376e5dc9be032e
exceptiongroup==1.2.1
executing==2.0.1
faiss-cpu==1.8.0
Faker==19.13.0
fastapi==0.110.2
fastjsonschema==2.19.1
filelock==3.13.4
flatbuffers==24.3.25
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.18.0
google-auth==2.29.0
google-cloud-aiplatform==1.48.0
google-cloud-bigquery==3.21.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.3
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
google_search_results==2.4.2
googleapis-common-protos==1.63.0
gql==3.5.0
graphql-core==3.2.3
greenlet==3.0.3
grpc-google-iam-v1==0.13.0
grpcio==1.62.2
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.25.2
huggingface-hub==0.22.2
humanfriendly==10.0
idna==3.7
immutables==0.20
importlib-metadata==7.0.0
importlib_resources==6.4.0
inflection==0.5.1
installer==0.7.0
ipykernel==6.29.4
ipython==8.18.0
ipywidgets==8.1.2
isodate==0.6.1
isoduration==20.11.0
jaraco.classes==3.4.0
jedi==0.19.1
jeepney==0.8.0
Jinja2==3.1.3
joblib==1.4.0
json5==0.9.25
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.1
jupyterlab_widgets==3.0.10
keyring==24.3.1
kubernetes==29.0.0
lamini==2.1.8
lamini-configuration==0.8.3
langchain==0.1.16
langchain-community==0.0.34
langchain-core==0.1.45
langchain-openai==0.1.3
langchain-pinecone==0.1.0
langchain-text-splitters==0.0.1
langchainhub==0.1.15
langcodes==3.3.0
langfuse==2.27.1
langkit==0.0.29
langsmith==0.1.49
Levenshtein==0.25.1
llama-index-core==0.10.36
llama-parse==0.4.2
llamaindex-py-client==0.1.19
log10-io==0.8.2
loguru==0.7.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistralai==0.1.8
mistune==3.0.2
mmh3==4.1.0
monotonic==1.6
more-itertools==10.2.0
mosaicml-cli==0.5.34
mpmath==1.3.0
msgpack==1.0.8
msrest==0.7.1
multidict==6.0.5
multiprocess==0.70.16
murmurhash==1.0.10
mypy==1.10.0
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
notebook==7.1.3
notebook_shim==0.2.4
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
onnxruntime==1.17.3
openai==1.23.2
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-asgi==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-util-http==0.45b0
orjson==3.10.1
overrides==7.7.0
packaging==23.2
pandas==2.2.2
pandocfilters==1.5.1
parea-ai==0.2.139
parso==0.8.4
pexpect==4.9.0
phonenumbers==8.13.35
pillow==10.3.0
pinecone-client==3.2.2
pkginfo==1.10.0
platformdirs==3.11.0
poetry==1.8.2
poetry-core==1.9.0
poetry-plugin-dotenv==0.6.33
poetry-plugin-export==1.7.1
portkey-ai==1.2.4
posthog==3.5.0
preshed==3.0.9
presidio-analyzer==2.2.352
presidio-anonymizer==2.2.352
prometheus_client==0.20.0
prompt-toolkit==3.0.36
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.0.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.0
pydantic_core==2.18.1
pydeck==0.8.1b0
Pygments==2.18.0
pymongo==4.6.3
pypdf==4.2.0
pyphen==0.15.0
PyPika==0.48.9
pyproject_hooks==1.0.0
pyreadline3==3.4.1
pysbd==0.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2024.1
pyupgrade==3.15.2
PyYAML==6.0.1
pyzmq==26.0.3
qtconsole==5.5.1
QtPy==2.4.1
questionary==2.0.1
rapidfuzz==3.8.1
referencing==0.34.0
regex==2024.4.16
requests==2.31.0
requests-file==2.0.0
requests-futures==1.0.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.13.0
SecretStorage==3.3.3
Send2Trash==1.8.3
sentence-transformers==2.3.1
sentencepiece==0.2.0
sentry-sdk==1.45.0
shapely==2.0.4
shellingham==1.5.4
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.29
srsly==2.4.8
sseclient-py==1.8.0
stack-data==0.6.3
starlette==0.37.2
streamlit==1.33.0
streamlit-feedback==0.1.3
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
terminado==0.18.1
textstat==0.7.3
thinc==8.2.3
threadpoolctl==3.4.0
tiktoken==0.6.0
tinycss2==1.3.0
tldextract==5.1.2
together==0.2.11
tokenize-rt==5.2.0
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.4
toolz==0.12.1
torch==2.2.2
tornado==6.4
tqdm==4.66.2
traitlets==5.14.3
transformers==4.40.0
triton==2.2.0
trove-classifiers==2024.4.10
typer==0.9.4
types-python-dateutil==2.9.0.20240316
types-requests==2.31.0.20240406
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
validators==0.28.1
virtualenv==20.26.0
wasabi==1.1.2
watchdog==4.0.0
watchfiles==0.21.0
wcwidth==0.2.13
weasel==0.3.4
webcolors==1.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==11.0.3
whylabs-client==0.6.3
whylogs==1.3.32
whylogs-sketching==3.4.1.dev3
widgetsnbextension==4.0.10
wikipedia==1.4.0
win32-setctime==1.1.0
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.18.1
